+++ components/userSetting/account.vue
@@ -1,0 +1,52 @@
+<script lang="ts" setup>
+const modalDeleteUser = ref<boolean>(false);
+</script>
+
+<template>
+    <div class="flex flex-col gap-3">
+        <div
+            class="gap-3 p-3 mt-6 rounded-lg flex flex-col border border-zinc-300 dark:border-zinc-400"
+        >
+            <UiTitle
+                label="アカウント操作"
+                icon="lucide:circle-alert"
+                is="h2"
+            />
+            <div class="w-full pl-3 flex gap-2 justify-between items-center">
+                <div class="flex flex-col gap-1">
+                    <h3 class="text-sm font-semibold">パスワード変更</h3>
+                </div>
+                <ButtonBase
+                    label="パスワード変更"
+                    class="text-red-500 dark:text-red-400 hover:text-white hover:dark:text-white hover:bg-red-500 hover:dark:bg-red-800 hover:outline-red-400 hover:dark:outline-red-700"
+                    @click="modalDeleteUser = true"
+                />
+            </div>
+        </div>
+
+        <div
+            class="gap-3 p-3 rounded-lg flex flex-col border border-red-300 dark:border-red-400"
+        >
+            <UiTitle label="DANGER ZONE" icon="lucide:circle-alert" is="h2" />
+            <div class="w-full pl-3 flex gap-2 justify-between items-center">
+                <div class="flex flex-col gap-1">
+                    <h3 class="text-sm font-semibold">アカウント削除</h3>
+                    <p class="text-xs">
+                        アカウントおよびアカウントに紐づくデータをすべて削除します。<br />
+                        削除したアカウントは復元できません。
+                    </p>
+                </div>
+                <ButtonBase
+                    label="アカウント削除"
+                    class="text-red-500 dark:text-red-400 hover:text-white hover:dark:text-white hover:bg-red-500 hover:dark:bg-red-800 hover:outline-red-400 hover:dark:outline-red-700"
+                    @click="modalDeleteUser = true"
+                />
+            </div>
+        </div>
+    </div>
+
+    <ModalDeleteUser
+        v-model="modalDeleteUser"
+        @close="modalDeleteUser = false"
+    />
+</template>
