+++ package.json
@@ -51,6 +51,7 @@
         "sharp": "^0.33.5",
         "tailwind-merge": "^2.6.0",
         "tailwind-variants": "^0.3.0",
+        "vue-draggable-plus": "^0.6.0",
         "zod": "^3.24.1"
     },
     "volta": {
